+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # CHANGELOG
 
-## 0.1.1 (unreleased)
+## 0.2.0 (unreleased)
 
 - add missing `OptionalClass` and `OptionalMetaclass` classes
+- rename `McsArgs.model_repr` to `McsArgs.repr`
+- rename `McsArgs.model_meta` to `McsArgs.meta`
+- rename the `base_model_meta` argument to `MetaOption.get_value` to `base_classes_meta`
 
 ## 0.1.0 (2018/09/24)